Split Avahi stuff into its own file.
